Finally uploaded my python codes.
dscrt_r-swap.py is still a work in progress, I couldn't quite get it to work the way I wanted. quat_math.py is essentially a library for doing quaternion and bi-quaternion operations, although it seems as though my biquat_sqrt function has an error in it. rand_rot.py is the code that shows the rotations given by the sqrt of a random quaternion yield the borne rule.